Move computations used by sumcheck
