+++ barretenberg/cpp/src/barretenberg/honk/proof_system/power_polynomial.test.cpp
