+++ barretenberg/cpp/src/barretenberg/proof_system/library/grand_product_delta.hpp
