+++ barretenberg/cpp/src/barretenberg/proof_system/library/grand_product_library.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include "barretenberg/common/constexpr_utils.hpp"
-#include "barretenberg/honk/sumcheck/sumcheck.hpp"
 #include "barretenberg/plonk/proof_system/proving_key/proving_key.hpp"
 #include "barretenberg/polynomials/polynomial.hpp"
 #include <typeinfo>
